Add filterIds to UpdateDataDestinationRequest
Ambar will now allow for swapping of filters in place on DataDestiantions. This
means filters can be replaced with a new filter which references the same
DataSource. Adding filters which refer to a new DataSource, or which refer to
an already referenced dataSource are not supported, nor is removing filters.

These types of mutations should instead delete and recreate the DataDestiantion.

Note: Message transmission is not reset when a filter is swapped, so ensure you
update your filters and backend systems in accordance with the result you expect.